feat: better express and pino typing
Latest pino version does not allow for custom typing of the LogFn
function, as such, the only thing we can do is assert type in the
AuditLog function.

Express now shadows the global type to extend User, this removes the
need to explicitly cast any Express request user as AuthUser.